refactor: parser -> converter로 전환
